requirements.txt
langchain 
langchain-ollama 
ollama
elasticsearch
elasticsearch-dsl
newspaper3k
lxml_html_clean
streamlit==1.21.0